requirements.txt
aiogram==3.13.1
openai==1.54.4
anthropic==0.40.0
python-docx==1.1.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
aiofiles==24.1.0